Create LCD_test.c
This is simply to test out the LCD display before we implement it onto the project itself, I want to make sure I know how it works before we start messing around with it. I'm using code from the book and translating it for our purposes, so far the basic functionality and setup are in, needs to be tested first.